Move meter value processing to base ChargePoint
